+++ app/core/contract/ConnectionInterface.php
@@ -14,4 +14,6 @@ interface ConnectionInterface
     public function beginTransaction(): bool;
     public function commit(): bool;
     public function rollback(): bool;
+    public function inTransaction(): bool;
+    public function getDatabase(): string;
 }